feat(dashboard-strategy-health-component): refactor of functions and types
