--- src/components/Report/BoardsOverview/queries.gql
@@ -1,14 +1,0 @@
-query GET_HEALTH_QUANTITIES {
-  me {
-    companies {
-      quantities {
-        keyResultsQuantity
-        objectivesQuantity
-        highConfidence
-        mediumConfidence
-        lowConfidence
-        barrier
-      }
-    }
-  }
-}
